Don't crash when receiving zulip requests in personals
Apparently zulip now reports the display_recipient as a list on personals,
which means it isn't hashable. However, we try to use the class as part of our
anti-duplication/rate-limiting dict's key. We add personals to that dict even
though we don't use them, which caused chiron to crash when trying to hash the
list. This changes personals to use None as the key, so it will no longer
crash.