Sort of attempt to pull PyZephyr from Github
PyPI's PyZephyr is too old, so we need to use the Github version. I tried a few
things, which didn't seem to work, but here's some of the detritus. It appears
that pip (or whatever) pretty reasonably doesn't want to pull packages from
arbitrary places, so we may be out of luck until fixing PyZephyr on PyPI.