Convert to CircleCI 2.0
CircleCI 1.0 is shutting down at the end of August 2018
(https://circleci.com/sunset1-0/), so it seems prudent to move off of it. It
appears that (for now, at least) the Python 3.5 image also has Python 2.7, and
I like using tox to drive our multi-version support, so I'm using a single
Docker image for both versions for now.